Add new files for backend application
Several new Python files were created, including main.py, that contain functions and classes for running a REST API server with FastAPI. Adapters and connectors to different databases and an Oracle subsystem are implemented. Schedulers, routers, and a test for routes are added as well. The code also introduces handling of .xlsx files, generation of PDF documents, and serialization of database responses. A Dockerfile and docker-compose were created to containerize the application. Corresponding configuration files, such as .gitignore, .sonarcloud.properties, and .dockerignore, are updated appropriately.